refactor(data): split queries.py 1283-line god module into domain package
Per CLAUDE.md (single file <300 lines + single responsibility), split
queries.py into 7 domain files; memory.py exceeded 300 lines so it was
sub-split into 3 (memory + memory_edges + memory_search) per spec §3.6:
  - model_provider.py (3 fns, 45 lines)
  - persona.py (6 fns, 89 lines)
  - messages.py (12 fns, 238 lines)
  - agent_response.py (4 fns, 103 lines, agent_responses table)
  - schedule.py (11 fns, 218 lines)
  - life.py (9 fns, 172 lines)
  - memory.py (13 fns, 198 lines, fragments + abstracts CRUD)
  - memory_edges.py (7 fns, 130 lines, edges + notes)
  - memory_search.py (9 fns, 222 lines, read helpers)

Domain assigned by primary table operated on (not by call site):
  - list_today_fragments lives in life module by usage but operates on
    fragments table -> memory_search.py
  - find_gray_config joins via message_id but reads LarkBaseChatInfo
    chat-dimension config -> messages.py (not persona)
  - is_chat_request_completed / set_agent_response_bot / *_safety_status
    all primarily on agent_responses -> agent_response.py (carved out
    from messages.py)

queries/__init__.py re-exports via from-X-import-*; __all__ aggregates
the 9 domain modules. queries.py file deleted (no compat shim).

Hard test test_queries_split.py asserts:
  1. queries.__all__ equals 74-function set from spec §3.3
  2. domain __all__ pairwise-disjoint (from-import-* name collision is
     silent; ruff/mypy do not catch it)
  3. each exported name is callable

Caller import sites unchanged: 0 caller files modified.

Spec: docs/superpowers/specs/2026-05-07-dataflow-phase-6-cleanup-design.md §3
