--- apps/agent-service/app/data/queries.py
@@ -1,1283 +1,0 @@
-"""Domain queries — pure data access, no business logic.
-
-All functions accept ``session: AsyncSession`` as the first parameter.
-Callers manage transactions via ``get_session()`` context manager.
-
-Sections:
-  - Model Provider
-  - Persona / Bot config
-  - Chat messages
-  - Schedule
-  - Life Engine
-  - Memory v4 — fragments, abstracts, edges, notes, glimpse, reply style
-"""
-
-from __future__ import annotations
-
-import json
-from datetime import UTC, datetime, timedelta, timezone
-
-from sqlalchemy import func, or_, text, update
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.future import select
-
-from app.data.models import (
-    AbstractMemory,
-    AkaoSchedule,
-    BotPersona,
-    ConversationMessage,
-    Fragment,
-    GlimpseState,
-    LarkBaseChatInfo,
-    LarkGroupChatInfo,
-    LarkGroupMember,
-    LarkUser,
-    LifeEngineState,
-    MemoryEdge,
-    ModelMapping,
-    ModelProvider,
-    Note,
-    ReplyStyleLog,
-    ScheduleRevision,
-)
-
-# CST timezone for date boundary calculations
-_CST = timezone(timedelta(hours=8))
-
-
-# --- Model Provider ---
-
-
-def parse_model_id(model_id: str) -> tuple[str, str]:
-    """Parse ``"provider:model"`` into ``(provider_name, model_name)``.
-
-    Falls back to ``"302.ai"`` as default provider when no colon present.
-    """
-    if ":" in model_id:
-        provider_name, model_name = model_id.split(":", 1)
-        return provider_name.strip(), model_name.strip()
-    return "302.ai", model_id.strip()
-
-
-async def find_model_mapping(session: AsyncSession, alias: str) -> ModelMapping | None:
-    """Look up a model mapping by alias."""
-    result = await session.execute(
-        select(ModelMapping).where(ModelMapping.alias == alias)
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_provider_by_name(
-    session: AsyncSession, name: str
-) -> ModelProvider | None:
-    """Look up a model provider by name."""
-    result = await session.execute(
-        select(ModelProvider).where(ModelProvider.name == name)
-    )
-    return result.scalar_one_or_none()
-
-
-# --- Persona / Bot config ---
-
-
-async def find_persona(session: AsyncSession, persona_id: str) -> BotPersona | None:
-    """Fetch a bot persona by primary key."""
-    return await session.get(BotPersona, persona_id)
-
-
-async def list_all_persona_ids(session: AsyncSession) -> list[str]:
-    """Return all persona_id values from bot_persona table."""
-    result = await session.execute(select(BotPersona.persona_id))
-    return [row[0] for row in result.all()]
-
-
-async def find_gray_config(session: AsyncSession, message_id: str) -> dict | None:
-    """Look up gray_config for the chat that a message belongs to."""
-    stmt = (
-        select(LarkBaseChatInfo.gray_config)
-        .join(
-            ConversationMessage,
-            ConversationMessage.chat_id == LarkBaseChatInfo.chat_id,
-        )
-        .where(ConversationMessage.message_id == message_id)
-    )
-    return await session.scalar(stmt)
-
-
-async def resolve_persona_id(session: AsyncSession, bot_name: str) -> str:
-    """Map bot_name -> persona_id via bot_config table. Falls back to bot_name itself."""
-    result = await session.execute(
-        text("SELECT persona_id FROM bot_config WHERE bot_name = :bn"),
-        {"bn": bot_name},
-    )
-    row = result.scalar_one_or_none()
-    return row if row else bot_name
-
-
-async def resolve_bot_name_for_persona(
-    session: AsyncSession, persona_id: str, chat_id: str
-) -> str | None:
-    """Find the bot_name that should send messages for a persona in a specific chat."""
-    result = await session.execute(
-        text(
-            "SELECT bc.bot_name FROM bot_config bc "
-            "JOIN bot_chat_presence bp ON bc.bot_name = bp.bot_name "
-            "WHERE bp.chat_id = :cid AND bp.is_active = true "
-            "AND bc.persona_id = :pid AND bc.is_active = true "
-            "LIMIT 1"
-        ),
-        {"cid": chat_id, "pid": persona_id},
-    )
-    return result.scalar_one_or_none()
-
-
-async def resolve_mentioned_personas(
-    session: AsyncSession, mentions: list[str]
-) -> list[str]:
-    """Map mention app_id list to persona_id list via bot_config table."""
-    result = await session.execute(
-        text(
-            "SELECT DISTINCT persona_id FROM bot_config "
-            "WHERE app_id = ANY(:mentions) "
-            "AND is_active = true "
-            "AND persona_id IS NOT NULL"
-        ),
-        {"mentions": mentions},
-    )
-    return [row[0] for row in result.fetchall()]
-
-
-async def find_bot_names_for_persona(
-    session: AsyncSession, persona_id: str
-) -> list[str]:
-    """Return all active bot_names mapped to a persona_id."""
-    result = await session.execute(
-        text(
-            "SELECT bot_name FROM bot_config "
-            "WHERE persona_id = :pid AND is_active = true"
-        ),
-        {"pid": persona_id},
-    )
-    return list(result.scalars().all())
-
-
-async def find_cross_chat_messages(
-    session: AsyncSession,
-    user_id: str,
-    bot_names: list[str],
-    exclude_chat_id: str,
-    since_ms: int,
-    excluded_chat_ids: list[str] | None = None,
-) -> list[ConversationMessage]:
-    """Fetch recent cross-chat interactions between a user and a persona.
-
-    Returns user messages + bot replies across all chat types (group + p2p).
-    Excludes the current chat and any blacklisted chat IDs.
-    """
-    stmt = (
-        select(ConversationMessage)
-        .where(ConversationMessage.chat_id != exclude_chat_id)
-        .where(ConversationMessage.create_time >= since_ms)
-        .where(ConversationMessage.bot_name.in_(bot_names))
-        .where(
-            or_(
-                # User's messages
-                (ConversationMessage.role == "user")
-                & (ConversationMessage.user_id == user_id),
-                # Bot's assistant replies
-                ConversationMessage.role == "assistant",
-            )
-        )
-        .order_by(ConversationMessage.create_time.asc())
-    )
-    if excluded_chat_ids:
-        stmt = stmt.where(~ConversationMessage.chat_id.in_(excluded_chat_ids))
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-# --- Chat messages ---
-
-
-async def find_message_content(session: AsyncSession, message_id: str) -> str | None:
-    """Fetch message content by message_id."""
-    stmt = select(ConversationMessage.content).where(
-        ConversationMessage.message_id == message_id
-    )
-    return await session.scalar(stmt)
-
-
-async def find_messages_in_range(
-    session: AsyncSession,
-    chat_id: str,
-    start_time: int,
-    end_time: int,
-    limit: int = 2000,
-) -> list[ConversationMessage]:
-    """Fetch messages in a chat within a time range (ascending)."""
-    result = await session.execute(
-        select(ConversationMessage)
-        .where(ConversationMessage.chat_id == chat_id)
-        .where(ConversationMessage.create_time >= start_time)
-        .where(ConversationMessage.create_time < end_time)
-        .order_by(ConversationMessage.create_time.asc())
-        .limit(limit)
-    )
-    return list(result.scalars().all())
-
-
-async def find_username(session: AsyncSession, user_id: str) -> str | None:
-    """Look up display name from lark_user by union_id."""
-    result = await session.execute(
-        select(LarkUser.name).where(LarkUser.union_id == user_id)
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_group_name(session: AsyncSession, chat_id: str) -> str | None:
-    """Look up group name from lark_group_chat_info."""
-    result = await session.execute(
-        select(LarkGroupChatInfo.name).where(LarkGroupChatInfo.chat_id == chat_id)
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_group_download_permission(
-    session: AsyncSession, chat_id: str
-) -> str | None:
-    """Fetch download_has_permission_setting for a group chat, or None."""
-    result = await session.execute(
-        select(LarkGroupChatInfo.download_has_permission_setting).where(
-            LarkGroupChatInfo.chat_id == chat_id
-        )
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_message_by_id(
-    session: AsyncSession, message_id: str
-) -> ConversationMessage | None:
-    """Fetch full message object by message_id."""
-    result = await session.execute(
-        select(ConversationMessage).where(ConversationMessage.message_id == message_id)
-    )
-    return result.scalar_one_or_none()
-
-
-async def resolve_message_id_by_row_id(
-    session: AsyncSession, row_id: str | int
-) -> str | None:
-    """Resolve int row id to lark message_id (om_xxx). Returns None if not found."""
-    try:
-        rid = int(row_id)
-    except (ValueError, TypeError):
-        return None
-    result = await session.execute(
-        select(ConversationMessage.message_id).where(ConversationMessage.id == rid)
-    )
-    return result.scalar_one_or_none()
-
-
-async def set_agent_response_bot(
-    session: AsyncSession,
-    session_id: str,
-    bot_name: str,
-    persona_id: str,
-) -> None:
-    """Update bot_name and persona_id on agent_responses row."""
-    await session.execute(
-        text(
-            "UPDATE agent_responses SET bot_name = :bn, persona_id = :pid "
-            "WHERE session_id = :sid"
-        ),
-        {"bn": bot_name, "pid": persona_id, "sid": session_id},
-    )
-
-
-async def is_chat_request_completed(
-    session: AsyncSession,
-    session_id: str | None,
-    *,
-    is_proactive: bool = False,
-) -> bool:
-    """Return whether a chat_request redelivery should be treated as done."""
-    if not session_id:
-        return False
-
-    if is_proactive:
-        result = await session.execute(
-            select(func.count())
-            .select_from(ConversationMessage)
-            .where(ConversationMessage.response_id == session_id)
-            .where(ConversationMessage.role == "assistant")
-        )
-        return (result.scalar_one() or 0) > 0
-
-    result = await session.execute(
-        text("SELECT status FROM agent_responses WHERE session_id = :sid"),
-        {"sid": session_id},
-    )
-    status = result.scalar_one_or_none()
-    return status in ("completed", "recalled")
-
-
-async def get_safety_status(
-    session: AsyncSession, session_id: str
-) -> str | None:
-    """Read ``safety_status`` from ``agent_responses``; None if row missing.
-
-    Phase 2 ``run_post_safety`` 节点入口判 None 时 raise（让 durable
-    handler 进 DLQ）—— None 不再被当成 fail-open 的 pending 处理，
-    见 spec §3.8 / §4.4。
-    """
-    result = await session.execute(
-        text("SELECT safety_status FROM agent_responses WHERE session_id = :sid"),
-        {"sid": session_id},
-    )
-    return result.scalar_one_or_none()
-
-
-async def set_safety_status(
-    session: AsyncSession,
-    session_id: str,
-    status: str,
-    result_json: dict | None = None,
-) -> None:
-    """Update safety_status (and optional result) on agent_responses row."""
-    await session.execute(
-        text(
-            "UPDATE agent_responses "
-            "SET safety_status = :status, "
-            "    safety_result = CAST(:result AS jsonb), "
-            "    updated_at = NOW() "
-            "WHERE session_id = :session_id"
-        ),
-        {
-            "status": status,
-            "result": (json.dumps(result_json) if result_json is not None else None),
-            "session_id": session_id,
-        },
-    )
-
-
-async def find_last_bot_reply_time(session: AsyncSession, chat_id: str) -> int:
-    """Return the latest assistant reply create_time (ms) in a chat, or 0."""
-    result = await session.execute(
-        select(func.max(ConversationMessage.create_time)).where(
-            ConversationMessage.chat_id == chat_id,
-            ConversationMessage.role == "assistant",
-        )
-    )
-    return result.scalar_one_or_none() or 0
-
-
-# --- Schedule ---
-
-
-async def find_active_schedules_for_date(
-    session: AsyncSession, now_date: str
-) -> list[AkaoSchedule]:
-    """Fetch all active schedule entries covering a given date.
-
-    Returns raw entries — time-slot matching is the caller's responsibility.
-    """
-    result = await session.execute(
-        select(AkaoSchedule)
-        .where(AkaoSchedule.is_active.is_(True))
-        .where(AkaoSchedule.period_start <= now_date)
-        .where(AkaoSchedule.period_end >= now_date)
-        .order_by(AkaoSchedule.plan_type.asc())
-    )
-    return list(result.scalars().all())
-
-
-async def find_latest_plan(
-    session: AsyncSession,
-    plan_type: str,
-    before_date: str,
-    persona_id: str,
-) -> AkaoSchedule | None:
-    """Find the most recent plan of a type ending before a given date."""
-    result = await session.execute(
-        select(AkaoSchedule)
-        .where(AkaoSchedule.plan_type == plan_type)
-        .where(AkaoSchedule.is_active.is_(True))
-        .where(AkaoSchedule.period_end < before_date)
-        .where(AkaoSchedule.persona_id == persona_id)
-        .order_by(AkaoSchedule.period_end.desc())
-        .limit(1)
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_plan_for_period(
-    session: AsyncSession,
-    plan_type: str,
-    period_start: str,
-    period_end: str,
-    persona_id: str,
-) -> AkaoSchedule | None:
-    """Look up a plan by exact period boundaries."""
-    result = await session.execute(
-        select(AkaoSchedule)
-        .where(AkaoSchedule.plan_type == plan_type)
-        .where(AkaoSchedule.period_start == period_start)
-        .where(AkaoSchedule.period_end == period_end)
-        .where(AkaoSchedule.persona_id == persona_id)
-    )
-    return result.scalar_one_or_none()
-
-
-async def find_daily_entries(
-    session: AsyncSession, target_date: str, persona_id: str
-) -> list[AkaoSchedule]:
-    """Fetch all daily time-slot entries for a given date."""
-    result = await session.execute(
-        select(AkaoSchedule)
-        .where(AkaoSchedule.plan_type == "daily")
-        .where(AkaoSchedule.period_start == target_date)
-        .where(AkaoSchedule.is_active.is_(True))
-        .where(AkaoSchedule.persona_id == persona_id)
-        .order_by(AkaoSchedule.time_start.asc())
-    )
-    return list(result.scalars().all())
-
-
-async def list_schedules(
-    session: AsyncSession,
-    *,
-    plan_type: str | None = None,
-    persona_id: str | None = None,
-    active_only: bool = True,
-    limit: int = 50,
-) -> list[AkaoSchedule]:
-    """List schedule entries with optional filters."""
-    stmt = select(AkaoSchedule)
-    if plan_type:
-        stmt = stmt.where(AkaoSchedule.plan_type == plan_type)
-    if persona_id:
-        stmt = stmt.where(AkaoSchedule.persona_id == persona_id)
-    if active_only:
-        stmt = stmt.where(AkaoSchedule.is_active.is_(True))
-    stmt = stmt.order_by(
-        AkaoSchedule.period_start.desc(), AkaoSchedule.time_start.asc()
-    ).limit(limit)
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-async def upsert_schedule(session: AsyncSession, entry: AkaoSchedule) -> AkaoSchedule:
-    """Insert or update a schedule entry (matched by unique constraint)."""
-    result = await session.execute(
-        select(AkaoSchedule)
-        .where(AkaoSchedule.persona_id == entry.persona_id)
-        .where(AkaoSchedule.plan_type == entry.plan_type)
-        .where(AkaoSchedule.period_start == entry.period_start)
-        .where(AkaoSchedule.period_end == entry.period_end)
-        .where(
-            AkaoSchedule.time_start == entry.time_start
-            if entry.time_start
-            else AkaoSchedule.time_start.is_(None)
-        )
-    )
-    existing = result.scalar_one_or_none()
-
-    if existing:
-        existing.content = entry.content
-        existing.mood = entry.mood
-        existing.energy_level = entry.energy_level
-        existing.response_style_hint = entry.response_style_hint
-        existing.proactive_action = entry.proactive_action
-        existing.target_chats = entry.target_chats
-        existing.model = entry.model
-        existing.is_active = entry.is_active
-        await session.flush()
-        await session.refresh(existing)
-        return existing
-
-    session.add(entry)
-    await session.flush()
-    await session.refresh(entry)
-    return entry
-
-
-async def delete_schedule(session: AsyncSession, schedule_id: int) -> bool:
-    """Delete a schedule entry by id. Returns True if found and deleted."""
-    result = await session.execute(
-        select(AkaoSchedule).where(AkaoSchedule.id == schedule_id)
-    )
-    entry = result.scalar_one_or_none()
-    if not entry:
-        return False
-    await session.delete(entry)
-    return True
-
-
-# --- Life Engine ---
-
-
-async def find_latest_life_state(
-    session: AsyncSession, persona_id: str
-) -> LifeEngineState | None:
-    """Fetch the most recent life engine state row."""
-    result = await session.execute(
-        select(LifeEngineState)
-        .where(LifeEngineState.persona_id == persona_id)
-        .order_by(LifeEngineState.created_at.desc())
-        .limit(1)
-    )
-    return result.scalar_one_or_none()
-
-
-async def insert_life_state(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    current_state: str,
-    activity_type: str,
-    response_mood: str,
-    skip_until: datetime | None,
-    reasoning: str | None = None,
-    state_end_at: datetime | None = None,
-) -> int:
-    """INSERT a new life engine state row (append-only). Returns the new row id."""
-    row = LifeEngineState(
-        persona_id=persona_id,
-        current_state=current_state,
-        activity_type=activity_type,
-        response_mood=response_mood,
-        reasoning=reasoning,
-        skip_until=skip_until,
-        state_end_at=state_end_at,
-    )
-    session.add(row)
-    await session.flush()
-    return row.id
-
-
-async def find_today_activity_states(
-    session: AsyncSession,
-    persona_id: str,
-    today_start: datetime,
-) -> list[LifeEngineState]:
-    """Fetch activity states created today (ascending)."""
-    result = await session.execute(
-        select(LifeEngineState)
-        .where(LifeEngineState.persona_id == persona_id)
-        .where(LifeEngineState.created_at >= today_start)
-        .order_by(LifeEngineState.created_at.asc())
-    )
-    return list(result.scalars().all())
-
-
-# --- Memory — fragments ---
-
-
-async def list_today_fragments(
-    session: AsyncSession,
-    persona_id: str,
-    *,
-    sources: list[str] | None = None,
-) -> list[Fragment]:
-    """Fetch v4 Fragments created today (CST 00:00+, ascending). Skips forgotten rows."""
-    today_cst = datetime.now(_CST).replace(hour=0, minute=0, second=0, microsecond=0)
-    stmt = (
-        select(Fragment)
-        .where(Fragment.persona_id == persona_id)
-        .where(Fragment.created_at >= today_cst)
-        .where(Fragment.clarity != "forgotten")
-    )
-    if sources:
-        stmt = stmt.where(Fragment.source.in_(sources))
-    stmt = stmt.order_by(Fragment.created_at.asc())
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-async def find_fragments_since(
-    session: AsyncSession,
-    persona_id: str,
-    since_dt: datetime,
-    *,
-    sources: list[str] | None = None,
-    limit: int = 50,
-) -> list[Fragment]:
-    """Fetch v4 Fragments created since a given datetime (descending). Skips forgotten rows."""
-    stmt = (
-        select(Fragment)
-        .where(Fragment.persona_id == persona_id)
-        .where(Fragment.created_at >= since_dt)
-        .where(Fragment.clarity != "forgotten")
-    )
-    if sources:
-        stmt = stmt.where(Fragment.source.in_(sources))
-    stmt = stmt.order_by(Fragment.created_at.desc()).limit(limit)
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-async def find_life_states_in_range(
-    session: AsyncSession,
-    persona_id: str,
-    start_dt: datetime,
-    end_dt: datetime,
-) -> list[LifeEngineState]:
-    """Fetch life_engine_state rows within a datetime range (ascending)."""
-    result = await session.execute(
-        select(LifeEngineState)
-        .where(LifeEngineState.persona_id == persona_id)
-        .where(LifeEngineState.created_at >= start_dt)
-        .where(LifeEngineState.created_at < end_dt)
-        .order_by(LifeEngineState.created_at.asc())
-    )
-    return list(result.scalars().all())
-
-
-# --- Memory — glimpse ---
-
-
-async def find_latest_glimpse_state(
-    session: AsyncSession,
-    persona_id: str,
-    chat_id: str,
-) -> GlimpseState | None:
-    """Fetch the most recent glimpse state for a persona+chat pair."""
-    result = await session.execute(
-        select(GlimpseState)
-        .where(GlimpseState.persona_id == persona_id)
-        .where(GlimpseState.chat_id == chat_id)
-        .order_by(GlimpseState.id.desc())
-        .limit(1)
-    )
-    return result.scalar_one_or_none()
-
-
-async def insert_glimpse_state(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    chat_id: str,
-    last_seen_msg_time: int,
-    observation: str,
-) -> None:
-    """INSERT a new glimpse observation row (append-only)."""
-    session.add(
-        GlimpseState(
-            persona_id=persona_id,
-            chat_id=chat_id,
-            last_seen_msg_time=last_seen_msg_time,
-            observation=observation,
-        )
-    )
-
-
-# --- Memory — reply style ---
-
-
-async def insert_reply_style(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    style_text: str,
-    source: str,
-    observation: str | None = None,
-) -> None:
-    """Append a reply style audit log entry."""
-    session.add(
-        ReplyStyleLog(
-            persona_id=persona_id,
-            style_text=style_text,
-            source=source,
-            observation=observation,
-        )
-    )
-
-
-async def find_latest_reply_style(session: AsyncSession, persona_id: str) -> str | None:
-    """Fetch the most recent reply style text for a persona."""
-    result = await session.execute(
-        select(ReplyStyleLog.style_text)
-        .where(ReplyStyleLog.persona_id == persona_id)
-        .order_by(ReplyStyleLog.created_at.desc())
-        .limit(1)
-    )
-    return result.scalar_one_or_none()
-
-
-# --- Chat history — context messages and group members ---
-
-
-async def find_context_messages_for_anchors(
-    session: AsyncSession,
-    chat_id: str,
-    anchor_message_ids: list[str],
-    anchor_timestamps: list[int],
-    anchor_root_ids: set[str],
-    context_window_ms: int = 300_000,
-) -> list[tuple[ConversationMessage, LarkUser]]:
-    """Find messages surrounding anchor points (for search_group_history).
-
-    Returns list of (ConversationMessage, LarkUser) tuples.
-    """
-    time_conditions = [
-        ConversationMessage.create_time.between(
-            ts - context_window_ms, ts + context_window_ms
-        )
-        for ts in anchor_timestamps
-        if ts
-    ]
-    or_conditions = [
-        *time_conditions,
-        ConversationMessage.message_id.in_(anchor_message_ids),
-    ]
-    if anchor_root_ids:
-        or_conditions.append(
-            ConversationMessage.root_message_id.in_(anchor_root_ids)
-        )
-
-    stmt = (
-        select(ConversationMessage, LarkUser)
-        .join(LarkUser, ConversationMessage.user_id == LarkUser.union_id)
-        .where(
-            ConversationMessage.chat_id == chat_id,
-            or_(*or_conditions),
-        )
-        .order_by(ConversationMessage.create_time.asc())
-    )
-    result = await session.execute(stmt)
-    return list(result.all())
-
-
-async def find_group_members(
-    session: AsyncSession,
-    chat_id: str,
-    role: str | None = None,
-) -> list[tuple[LarkGroupMember, LarkUser]]:
-    """Find group members with user info.
-
-    Returns list of (LarkGroupMember, LarkUser) tuples.
-    """
-    stmt = (
-        select(LarkGroupMember, LarkUser)
-        .join(LarkUser, LarkGroupMember.union_id == LarkUser.union_id)
-        .where(
-            LarkGroupMember.chat_id == chat_id,
-            ~LarkGroupMember.is_leave,
-        )
-    )
-    if role == "owner":
-        stmt = stmt.where(LarkGroupMember.is_owner)
-    elif role == "manager":
-        stmt = stmt.where(LarkGroupMember.is_manager)
-
-    result = await session.execute(stmt)
-    return list(result.all())
-
-
-# --- Memory v4 — fragment / abstract_memory lookup ---
-
-
-async def get_fragment_by_id(
-    session: AsyncSession, fragment_id: str
-) -> Fragment | None:
-    """Fetch a v4 Fragment by primary key."""
-    result = await session.execute(select(Fragment).where(Fragment.id == fragment_id))
-    return result.scalar_one_or_none()
-
-
-async def get_abstract_by_id(
-    session: AsyncSession, abstract_id: str
-) -> AbstractMemory | None:
-    """Fetch a v4 AbstractMemory by primary key."""
-    result = await session.execute(
-        select(AbstractMemory).where(AbstractMemory.id == abstract_id)
-    )
-    return result.scalar_one_or_none()
-
-
-# --- Memory v4 — CRUD ---
-
-
-async def insert_fragment(
-    session: AsyncSession,
-    *,
-    id: str,
-    persona_id: str,
-    content: str,
-    source: str,
-    chat_id: str | None = None,
-    clarity: str = "clear",
-    created_at: datetime | None = None,
-) -> None:
-    f = Fragment(
-        id=id,
-        persona_id=persona_id,
-        content=content,
-        source=source,
-        chat_id=chat_id,
-        clarity=clarity,
-    )
-    if created_at is not None:
-        f.created_at = created_at
-        f.last_touched_at = created_at
-    session.add(f)
-
-
-async def touch_fragment(session: AsyncSession, fragment_id: str) -> None:
-    await session.execute(
-        update(Fragment)
-        .where(Fragment.id == fragment_id)
-        .values(last_touched_at=func.now())
-    )
-
-
-async def get_fragments_by_ids(
-    session: AsyncSession, ids: list[str]
-) -> list[Fragment]:
-    """Batch fetch fragments by id list. Preserves input order is NOT guaranteed."""
-    if not ids:
-        return []
-    result = await session.execute(
-        select(Fragment).where(Fragment.id.in_(ids))
-    )
-    return list(result.scalars().all())
-
-
-async def touch_fragments_bulk(session: AsyncSession, ids: list[str]) -> None:
-    """Update last_touched_at=NOW() for many fragments at once."""
-    if not ids:
-        return
-    await session.execute(
-        update(Fragment)
-        .where(Fragment.id.in_(ids))
-        .values(last_touched_at=func.now())
-    )
-
-
-async def insert_abstract_memory(
-    session: AsyncSession,
-    *,
-    id: str,
-    persona_id: str,
-    subject: str,
-    content: str,
-    created_by: str,
-    clarity: str = "clear",
-) -> None:
-    a = AbstractMemory(
-        id=id,
-        persona_id=persona_id,
-        subject=subject,
-        content=content,
-        created_by=created_by,
-        clarity=clarity,
-    )
-    session.add(a)
-
-
-async def touch_abstract(session: AsyncSession, abstract_id: str) -> None:
-    await session.execute(
-        update(AbstractMemory)
-        .where(AbstractMemory.id == abstract_id)
-        .values(last_touched_at=func.now())
-    )
-
-
-async def touch_abstracts_bulk(session: AsyncSession, ids: list[str]) -> None:
-    """Update last_touched_at=NOW() for many abstracts at once."""
-    if not ids:
-        return
-    await session.execute(
-        update(AbstractMemory)
-        .where(AbstractMemory.id.in_(ids))
-        .values(last_touched_at=func.now())
-    )
-
-
-async def count_abstracts_by_persona(
-    session: AsyncSession, persona_id: str
-) -> int:
-    result = await session.execute(
-        select(func.count())
-        .select_from(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-    )
-    return int(result.scalar_one())
-
-
-async def insert_memory_edge(
-    session: AsyncSession,
-    *,
-    id: str,
-    persona_id: str,
-    from_id: str,
-    from_type: str,
-    to_id: str,
-    to_type: str,
-    edge_type: str,
-    created_by: str,
-    reason: str | None = None,
-) -> None:
-    e = MemoryEdge(
-        id=id,
-        persona_id=persona_id,
-        from_id=from_id,
-        from_type=from_type,
-        to_id=to_id,
-        to_type=to_type,
-        edge_type=edge_type,
-        created_by=created_by,
-        reason=reason,
-    )
-    session.add(e)
-
-
-async def insert_note(
-    session: AsyncSession,
-    *,
-    id: str,
-    persona_id: str,
-    content: str,
-    when_at: datetime | None = None,
-) -> None:
-    n = Note(id=id, persona_id=persona_id, content=content, when_at=when_at)
-    session.add(n)
-
-
-async def get_active_notes(
-    session: AsyncSession, persona_id: str
-) -> list[Note]:
-    result = await session.execute(
-        select(Note)
-        .where(Note.persona_id == persona_id)
-        .where(Note.resolved_at.is_(None))
-        .order_by(Note.created_at.desc())
-    )
-    return list(result.scalars().all())
-
-
-async def resolve_note(
-    session: AsyncSession, *, note_id: str, resolution: str
-) -> None:
-    await session.execute(
-        update(Note)
-        .where(Note.id == note_id)
-        .values(resolved_at=func.now(), resolution=resolution)
-    )
-
-
-async def insert_schedule_revision(
-    session: AsyncSession,
-    *,
-    id: str,
-    persona_id: str,
-    content: str,
-    reason: str,
-    created_by: str,
-) -> None:
-    sr = ScheduleRevision(
-        id=id,
-        persona_id=persona_id,
-        content=content,
-        reason=reason,
-        created_by=created_by,
-    )
-    session.add(sr)
-
-
-async def get_current_schedule(
-    session: AsyncSession, persona_id: str
-) -> ScheduleRevision | None:
-    result = await session.execute(
-        select(ScheduleRevision)
-        .where(ScheduleRevision.persona_id == persona_id)
-        .order_by(ScheduleRevision.created_at.desc())
-        .limit(1)
-    )
-    return result.scalar_one_or_none()
-
-
-# --- Memory v4 — reviewer mutations ---
-
-
-async def update_abstract_content_query(
-    session: AsyncSession, *, abstract_id: str, new_content: str
-) -> None:
-    await session.execute(
-        update(AbstractMemory)
-        .where(AbstractMemory.id == abstract_id)
-        .values(content=new_content, last_touched_at=func.now())
-    )
-
-
-async def set_clarity(
-    session: AsyncSession, *, node_id: str, node_type: str, clarity: str
-) -> None:
-    if node_type == "abstract":
-        await session.execute(
-            update(AbstractMemory)
-            .where(AbstractMemory.id == node_id)
-            .values(clarity=clarity, last_touched_at=func.now())
-        )
-    elif node_type == "fact":
-        await session.execute(
-            update(Fragment)
-            .where(Fragment.id == node_id)
-            .values(clarity=clarity, last_touched_at=func.now())
-        )
-    else:
-        raise ValueError(f"unknown node_type {node_type}")
-
-
-async def delete_fragment_query(
-    session: AsyncSession, *, fragment_id: str
-) -> None:
-    # cascade delete edges touching this fragment first
-    await session.execute(
-        MemoryEdge.__table__.delete().where(
-            or_(MemoryEdge.from_id == fragment_id, MemoryEdge.to_id == fragment_id)
-        )
-    )
-    await session.execute(
-        Fragment.__table__.delete().where(Fragment.id == fragment_id)
-    )
-
-
-async def delete_edge(
-    session: AsyncSession, *, edge_id: str
-) -> None:
-    await session.execute(
-        MemoryEdge.__table__.delete().where(MemoryEdge.id == edge_id)
-    )
-
-
-async def list_fragments_window(
-    session: AsyncSession, *, persona_id: str, since: datetime,
-) -> list[Fragment]:
-    result = await session.execute(
-        select(Fragment)
-        .where(Fragment.persona_id == persona_id)
-        .where(Fragment.created_at >= since)
-        .where(Fragment.clarity != "forgotten")
-        .order_by(Fragment.created_at)
-    )
-    return list(result.scalars().all())
-
-
-async def list_abstracts_window(
-    session: AsyncSession, *, persona_id: str, since: datetime,
-) -> list[AbstractMemory]:
-    result = await session.execute(
-        select(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-        .where(AbstractMemory.created_at >= since)
-        .where(AbstractMemory.clarity != "forgotten")
-        .order_by(AbstractMemory.created_at)
-    )
-    return list(result.scalars().all())
-
-
-# --- Memory v4 — graph traversal ---
-
-
-async def list_edges_to(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    to_id: str,
-    edge_type: str | None = None,
-) -> list[MemoryEdge]:
-    """List edges whose ``to_id`` matches, optionally filtered by edge_type."""
-    stmt = (
-        select(MemoryEdge)
-        .where(MemoryEdge.persona_id == persona_id)
-        .where(MemoryEdge.to_id == to_id)
-    )
-    if edge_type:
-        stmt = stmt.where(MemoryEdge.edge_type == edge_type)
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-async def list_edges_from(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    from_id: str,
-    edge_type: str | None = None,
-) -> list[MemoryEdge]:
-    """List edges whose ``from_id`` matches, optionally filtered by edge_type."""
-    stmt = (
-        select(MemoryEdge)
-        .where(MemoryEdge.persona_id == persona_id)
-        .where(MemoryEdge.from_id == from_id)
-    )
-    if edge_type:
-        stmt = stmt.where(MemoryEdge.edge_type == edge_type)
-    result = await session.execute(stmt)
-    return list(result.scalars().all())
-
-
-async def get_abstracts_by_subject(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    subject: str,
-    limit: int = 20,
-) -> list[AbstractMemory]:
-    """Fetch non-forgotten abstracts for a subject, newest-touched first."""
-    result = await session.execute(
-        select(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-        .where(AbstractMemory.subject == subject)
-        .where(AbstractMemory.clarity != "forgotten")
-        .order_by(AbstractMemory.last_touched_at.desc())
-        .limit(limit)
-    )
-    return list(result.scalars().all())
-
-
-# --- Context injection helpers (Plan C) ---
-
-
-async def get_abstracts_by_subjects(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    subjects: list[str],
-    limit_per_subject: int = 5,
-) -> list[AbstractMemory]:
-    """Get abstracts whose subject is in given list (for always-on injection)."""
-    if not subjects:
-        return []
-    result = await session.execute(
-        select(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-        .where(AbstractMemory.subject.in_(subjects))
-        .where(AbstractMemory.clarity != "forgotten")
-        .order_by(
-            AbstractMemory.subject,
-            AbstractMemory.last_touched_at.desc(),
-        )
-    )
-    rows = list(result.scalars().all())
-    # Keep at most `limit_per_subject` per subject
-    by_subject: dict[str, list[AbstractMemory]] = {}
-    for r in rows:
-        by_subject.setdefault(r.subject, []).append(r)
-    out: list[AbstractMemory] = []
-    for subj in subjects:
-        out.extend(by_subject.get(subj, [])[:limit_per_subject])
-    return out
-
-
-async def get_recent_abstract_titles(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    limit: int = 10,
-) -> list[AbstractMemory]:
-    """Recently touched abstracts — for recall-index hint."""
-    result = await session.execute(
-        select(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-        .where(AbstractMemory.clarity != "forgotten")
-        .order_by(AbstractMemory.last_touched_at.desc())
-        .limit(limit)
-    )
-    return list(result.scalars().all())
-
-
-async def count_abstracts_per_subject_prefix(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    prefix: str,
-) -> int:
-    """Count non-forgotten abstracts whose subject starts with prefix."""
-    result = await session.execute(
-        select(func.count())
-        .select_from(AbstractMemory)
-        .where(AbstractMemory.persona_id == persona_id)
-        .where(AbstractMemory.subject.like(f"{prefix}%"))
-        .where(AbstractMemory.clarity != "forgotten")
-    )
-    return int(result.scalar_one())
-
-
-async def get_recent_fragments_for_injection(
-    session: AsyncSession,
-    *,
-    persona_id: str,
-    chat_id: str | None,
-    _trigger_user_id: str | None,
-    max_same_chat: int = 1,
-    max_other_chat: int = 2,
-    hours: int = 4,
-) -> list[Fragment]:
-    """短期注入规则：
-    - 当前 chat 最近 N 小时内的最新 1 条 fragment
-    - 其他 chat 最近 N 小时内的 fragment（最多 max_other_chat 条，每 chat 只取最新）
-
-    ``_trigger_user_id``: reserved for future user-scoped filtering; currently ignored.
-    """
-    since = datetime.now(UTC) - timedelta(hours=hours)
-    stmt = (
-        select(Fragment)
-        .where(Fragment.persona_id == persona_id)
-        .where(Fragment.clarity != "forgotten")
-        .where(Fragment.created_at >= since)
-        .order_by(Fragment.created_at.desc())
-        .limit(max_same_chat + max_other_chat * 20)
-    )
-    result = await session.execute(stmt)
-    all_recent = list(result.scalars().all())
-
-    same_chat: list[Fragment] = []
-    other_chats: dict[str, Fragment] = {}
-    for f in all_recent:
-        if chat_id and f.chat_id == chat_id:
-            if len(same_chat) < max_same_chat:
-                same_chat.append(f)
-        elif f.chat_id and f.chat_id not in other_chats:
-            other_chats[f.chat_id] = f
-
-    other_list = list(other_chats.values())[:max_other_chat]
-    return same_chat + other_list
-
-
-# --- Schedule ---
-
-
-async def get_schedule_revision_by_id(
-    session: AsyncSession, revision_id: str
-) -> ScheduleRevision | None:
-    """Fetch a schedule_revision by id."""
-    result = await session.execute(
-        select(ScheduleRevision).where(ScheduleRevision.id == revision_id)
-    )
-    return result.scalar_one_or_none()
-
-
-async def list_recent_life_states(
-    session: AsyncSession, *, persona_id: str, since: datetime
-) -> list[LifeEngineState]:
-    result = await session.execute(
-        select(LifeEngineState)
-        .where(LifeEngineState.persona_id == persona_id)
-        .where(LifeEngineState.created_at >= since)
-        .order_by(LifeEngineState.created_at)
-    )
-    return list(result.scalars().all())
-
-
-async def list_recent_schedule_revisions(
-    session: AsyncSession, *, persona_id: str, since: datetime
-) -> list[ScheduleRevision]:
-    result = await session.execute(
-        select(ScheduleRevision)
-        .where(ScheduleRevision.persona_id == persona_id)
-        .where(ScheduleRevision.created_at >= since)
-        .order_by(ScheduleRevision.created_at)
-    )
-    return list(result.scalars().all())
